Add Initial source files
This patch adds initial structure for solidity files development.

Before this patch, in order to deploy the contract, a developer had
to run `npm deploy` and then copy the address and the ABI to the
profile.js file and export both of them.
The deploy.js compiled and deployed each time `npm deploy` was run.
From now, a developer can just run `npm compile` to actually compile,
where compiling the project will place the output in build directory
and `npm deploy` would deploy the contract, and the user have to copy
the address only to profile.js.
The ABI is compiled and parsed automatically.