Remove RequestType Enum
This patch removes RequestType Enum as it is not needed.
On the beginning, we thought that we will have 2 types of exchanges:
	ExchangeType.Request
	ExchangeType.Response
But as we understood that Exchanges are not necessary because we only
need the "Request" type, we changed the Exchange to Request, and
on this patch we deleted the RequestType Enum from the code and from
the tests.