Add debt rotation
This patch adds the debt rotation feature.
Debt rotation is our most complicated and our main or core feature.
It allows a mediator (a person who owes money and being owed money)
to "rotate" the debts (giving up on some amount of money/debt in
exchange of get the same amount to bedecreased for their debt).

E.g. Debtor Owes 10 to Mediator who owes 10 to Creditor.
Rotating 10 units will lead to a new situation which
will be: Debtor owes 10 to Creditor. Mediator is out of the equation.

Debt rotation would lead to less physical money transactions and
hence to save money for all actors in system who take part regularly.